+++ tortoise/filters.py
@@ -1,6 +1,6 @@
 import operator
 from functools import partial
-from typing import TYPE_CHECKING, Any, Dict, Iterable, Optional
+from typing import TYPE_CHECKING, Any, Dict, Iterable, Optional, Tuple
 
 from pypika import Table
 from pypika.enums import SqlTypes
@@ -51,6 +51,10 @@ def is_in(field: Term, value: Any) -> Criterion:
 
 def not_in(field: Term, value: Any) -> Criterion:
     return field.notin(value) | field.isnull()
+
+
+def between_and(field: Term, value: Tuple[Any, Any]) -> Criterion:
+    return field.between(value[0], value[1])
 
 
 def not_equal(field: Term, value: Any) -> Criterion:
@@ -234,6 +238,12 @@ def get_filters_for_field(
             "field": actual_field_name,
             "source_field": source_field,
             "operator": operator.lt,
+        },
+        f"{field_name}__range": {
+            "field": actual_field_name,
+            "source_field": source_field,
+            "operator": between_and,
+            "value_encoder": list_encoder,
         },
         f"{field_name}__contains": {
             "field": actual_field_name,